No llamo a delete_metadata si no existen fields asociados

+++ series_tiempo_ar_api/libs/custom_admins/admin.py
@@ -39,6 +39,8 @@ class CustomDistributionAdmin(DistributionAdmin):
 
     def delete_model(self, _, queryset):
         fields = Field.objects.filter(distribution__identifier__in=queryset.values_list('identifier', flat=True))
+        if not fields:
+            return
         delete_metadata(list(fields))
         queryset.delete()
